[fix] Remove event card duplications
